refactor(evm): compile the fusion scan out of the zkEVM build entirely
Replaces the constant flag with the build's own split: the scan, the fused
instruction and the per-frame site table all move to *.std.cs or behind #if, so
the zkEVM build drops them before compiling rather than relying on the compiler
and ILC to remove a constant false. Nothing about fusion survives into the guest
— not the type, not the CodeInfo fields, not the branch on the MSTORE path.

Verified against the built assemblies: MappingSlotFusion,
InstructionMappingSlotKeccak and FusionSites appear in the standard
Nethermind.Evm.dll and none appear in the zkEVM one.
